+++ src/psolver.hpp
@@ -59,7 +59,7 @@ namespace std
 class parallelSolver
 {
 private:
-  static constexpr unsigned int DIM = 2;
+  static constexpr unsigned short DIM = 2;
   //first of all, the mesh that will be saved as a member of solver,
   Mesh<DIM> _mesh;
   // a fe class solver object, to deal with the single elements of the mesh member
